Add files via upload
Started a logistic regression model and printed out an accuracy score along with its confusion matrix. It has an accuracy of 0.84, which isn't bad, but can be improved. Next time, I plan on removing outliers and leverage points, and figure out what to do with all of the NaNs so I can use more data instead of dropping it.